Checking pauseWriterThreshold on sending to avoid freezing the transport
