+++ src/Lime.Transport.Tcp.UnitTests/PipeServerTcpTransportTests.cs
@@ -5,6 +5,7 @@ using Shouldly;
 using System;
 using System.Buffers;
 using System.Reflection;
+using System.Threading;
 using System.Threading.Tasks;
 using Lime.Protocol.Serialization;
 using Lime.Protocol.Server;
@@ -19,16 +20,13 @@ namespace Lime.Transport.Tcp.UnitTests
     [TestFixture]    
     public class PipeServerTcpTransportTests : ServerTransportTestsBase
     {
-        public int BufferSize { get; set; }
-
-        public int MaxBufferSize { get; set; }
+        public int PauseWriterThreshold { get; set; }
 
         public MemoryPool<byte> MemoryPool { get; } = MemoryPool<byte>.Shared;
         
         protected override Task SetUpImpl()
         {
-            BufferSize = TcpTransport.DEFAULT_BUFFER_SIZE;
-            MaxBufferSize = TcpTransport.DEFAULT_MAX_BUFFER_SIZE;
+            PauseWriterThreshold = PipeTcpTransport.DEFAULT_PAUSE_WRITER_THRESHOLD;
             
             return base.SetUpImpl();
         }
@@ -37,195 +35,33 @@ namespace Lime.Transport.Tcp.UnitTests
 
         protected override ITransportListener CreateTransportListener(Uri uri, IEnvelopeSerializer envelopeSerializer)
         {
-            return new PipeTcpTransportListener(uri, null, envelopeSerializer, MaxBufferSize, MemoryPool, TraceWriter.Object);
+            return new PipeTcpTransportListener(uri, null, envelopeSerializer, PauseWriterThreshold, MemoryPool, TraceWriter.Object);
         }
 
         protected override ITransport CreateClientTransport(IEnvelopeSerializer envelopeSerializer)
         {
-            return new PipeTcpTransport(envelopeSerializer);
-        }        
+            return new PipeTcpTransport(envelopeSerializer, pauseWriterThreshold: PauseWriterThreshold);
+        }
 
         [Test]
-        public async Task ReceiveAsync_BiggerThanBufferSizeMessageEnvelope_ServerShouldReceive()
+        public async Task SendAsync_BiggerThanPauseWriterThresholdMessageEnvelope_TransportShouldBeClosed()
         {
             // Arrange
-            BufferSize = 64;
-            MaxBufferSize = BufferSize * 6;
-            var message = Dummy.CreateMessage(Dummy.CreateRandomString(BufferSize * 4));
+            PauseWriterThreshold = 1024;
+            var message = Dummy.CreateMessage(Dummy.CreateRandomString(PauseWriterThreshold * 3));
             var target = await GetTargetAsync();
 
             // Act
-            await ClientTransport.SendAsync(message, CancellationToken);
-            var actual = await target.ReceiveAsync(CancellationToken);
-
-            // Assert
-            actual.ShouldNotBeNull();
-            var actualMessage = actual.ShouldBeOfType<Message>();
-            CompareMessages(message, actualMessage);
-//            var jsonBuffer = GetJsonBuffer(target);
-//            jsonBuffer.Buffer.Length.ShouldBe(BufferSize);
-
-        }
-
-        [Test]
-        public async Task ReceiveAsync_MultipleBiggerThanBufferSizeMessageEnvelope_ServerShouldReceive()
-        {
-            // Arrange            
-            BufferSize = 64;
-            MaxBufferSize = BufferSize * 6;
-            var count = Dummy.CreateRandomInt(100);
-            var target = await GetTargetAsync();
-            var messages = new Message[count];
-
-            for (int i = 0; i < count; i++)
-            {
-                messages[i] = Dummy.CreateMessage(Dummy.CreateRandomString(BufferSize * (3 + i % 2)));
-            }
-            var actuals = new Envelope[count];
-
-
-            // Act
-            for (int i = 0; i < count; i++)
-            {
-                await ClientTransport.SendAsync(messages[i], CancellationToken);
-                actuals[i] = await target.ReceiveAsync(CancellationToken);
-            }
-
-            // Assert
-            for (int i = 0; i < count; i++)
-            {
-                var actual = actuals[i];
-                var message = messages[i];
-                actual.ShouldNotBeNull();
-                var actualMessage = actual.ShouldBeOfType<Message>();
-                CompareMessages(message, actualMessage);
-            }
-
-//            var jsonBuffer = GetJsonBuffer(target);
-//            jsonBuffer.Buffer.Length.ShouldBe(BufferSize);
-        }
-
-
-        [Test]
-        public async Task ReceiveAsync_DoubleWritesAndReads_ServerShouldReceive()
-        {
-            // Arrange            
-            BufferSize = 64;
-            MaxBufferSize = BufferSize * 12;
-            var count = 50;
-            var target = await GetTargetAsync();
-            var messages = new Message[count];
-
-            for (int i = 0; i < count; i++)
-            {
-                messages[i] = Dummy.CreateMessage(Dummy.CreateRandomString(BufferSize * 3));
-            }
-            var actuals = new Envelope[count];
-
-
-            // Act
-            for (int i = 0; i < count; i += 2)
-            {
-                await ClientTransport.SendAsync(messages[i], CancellationToken);
-                await ClientTransport.SendAsync(messages[i + 1], CancellationToken);
-                actuals[i] = await target.ReceiveAsync(CancellationToken);
-                actuals[i + 1] = await target.ReceiveAsync(CancellationToken);
-            }
-
-            // Assert
-            for (int i = 0; i < count; i++)
-            {
-                var actual = actuals[i];
-                var message = messages[i];
-                actual.ShouldNotBeNull();
-                var actualMessage = actual.ShouldBeOfType<Message>();
-                CompareMessages(message, actualMessage);
-            }
-//
-//            var jsonBuffer = GetJsonBuffer(target);
-//            jsonBuffer.Buffer.Length.ShouldBe(BufferSize);
-        }
-
-        [Test]
-        public async Task ReceiveAsync_InterleavedWritesAndReads_ServerShouldReceive()
-        {
-            // Arrange            
-            BufferSize = 64;
-            MaxBufferSize = BufferSize * 24;
-            var pageSize = 5;
-            var count = 9 * pageSize;
-            var target = await GetTargetAsync();
-            var messages = new Message[count];
-
-            for (int i = 0; i < count; i++)
-            {
-                messages[i] = Dummy.CreateMessage(Dummy.CreateRandomString(BufferSize * 3));
-            }
-            var actuals = new Envelope[count];
-
-
-            // Act
-            for (int i = 0; i < count; i += pageSize)
-            {
-                await ClientTransport.SendAsync(messages[i], CancellationToken);
-                await ClientTransport.SendAsync(messages[i + 1], CancellationToken);
-                await ClientTransport.SendAsync(messages[i + 2], CancellationToken);
-                actuals[i] = await target.ReceiveAsync(CancellationToken);
-                actuals[i + 1] = await target.ReceiveAsync(CancellationToken);
-                await ClientTransport.SendAsync(messages[i + 3], CancellationToken);
-                actuals[i + 2] = await target.ReceiveAsync(CancellationToken);
-                await ClientTransport.SendAsync(messages[i + 4], CancellationToken);
-                actuals[i + 3] = await target.ReceiveAsync(CancellationToken);
-                actuals[i + 4] = await target.ReceiveAsync(CancellationToken);
-            }
-
-            // Assert
-            for (int i = 0; i < count; i++)
-            {
-                var actual = actuals[i];
-                var message = messages[i];
-                actual.ShouldNotBeNull();
-                var actualMessage = actual.ShouldBeOfType<Message>();
-                CompareMessages(message, actualMessage);
-            }
-//
-//            var jsonBuffer = GetJsonBuffer(target);
-//            jsonBuffer.Buffer.Length.ShouldBe(BufferSize);
-        }
-
-        [Test]
-        public async Task ReceiveAsync_BiggerThanMaxBufferSizeMessageEnvelope_ServerShouldThrowBufferOverflowException()
-        {
-            // Arrange
-            BufferSize = 64;
-            MaxBufferSize = BufferSize * 4;
-            var message = Dummy.CreateMessage(Dummy.CreateRandomString(BufferSize * 5));
-            var target = await GetTargetAsync();
-
-            // Act
-            await ClientTransport.SendAsync(message, CancellationToken);
             try
             {
-                var actual = await target.ReceiveAsync(CancellationToken);
-                throw new Exception("The expected exception didn't occur");
+                await target.SendAsync(message, CancellationToken);
             }
-            catch (BufferOverflowException)
+            catch (Exception ex)
             {
-                target.IsConnected.ShouldBe(false);
+                var argumentException = ex.ShouldBeOfType<ArgumentException>();
+                argumentException.ParamName.ShouldBe("envelope");
+                target.IsConnected.ShouldBeFalse();
             }
         }
-
-        private static void CompareMessages(Message message, Message actualMessage)
-        {
-            actualMessage.Id.ShouldBe(message.Id);
-            actualMessage.From.ShouldBe(message.From);
-            actualMessage.To.ShouldBe(message.To);
-            actualMessage.Pp.ShouldBe(message.Pp);
-            actualMessage.Metadata.ShouldBe(message.Metadata);
-            actualMessage.Type.ShouldBe(message.Type);
-            actualMessage.Content.ToString().ShouldBe(message.Content.ToString());
-            actualMessage.Metadata.ShouldBe(message.Metadata);
-        }
-        
     }
 }
